pyproject.toml
[tool.poetry]
name = "graphql_tutorial"
version = "0.1.0"
description = ""
authors = ["Thomas van ERven"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
graphene-django = "^2.15.0"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.20.0"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
